moved help page to pages directory
